Move internal.vm_setup to its own module
